refactor(tools): rewrite all tools to tool() builder pattern
Replace handler-factory pattern with single-file tool() builder
definitions. Zod input/output schemas, format functions, and
annotations for all 7 PubMed tools.
